create pkg

+++ db/migrations/20240112092751_create_table_alerts.up.sql
@@ -4,10 +4,10 @@ CREATE TABLE todos
     background_color       VARCHAR(7) NOT NULL DEFAULT #22c55e,
     highlight_color       VARCHAR(7) NOT NULL DEFAULT #9333ea,
     text_color       VARCHAR(7) NOT NULL DEFAULT #ffffff,
-    duration    INT NOT NULL DEFAULT 5000,
+    duration    SMALLINT NOT NULL DEFAULT 5000,
     is_border BOOLEAN NOT NULL DEFAULT true,
     is_unit BOOLEAN NOT NULL DEFAULT false,
-    created_at  bigint    NOT NULL,
-    updated_at  bigint    NOT NULL,
+    created_at  BIGINT    NOT NULL,
+    updated_at  BIGINT    NOT NULL,
     PRIMARY KEY (id)
 ) engine = innodb;
